servers/yandexart/func/requirements.txt
yandex-ai-studio-sdk
Pillow
boto3>=1.34.0
